refactor(clients): relocate Claude/Codex loc-agent baselines to codeminer/clients/ (refs #150, #156)
Cherry-picked the third-party end-to-end localization baselines from #156
(Claude SDK + Codex SDK agents + dataset runner, authored by Hengjia Yu) and
moved the two agent clients out of codeminer/agent/ into a dedicated
codeminer/clients/ package (they are external-SDK API clients, not part of our
skill/runner agent stack). codeminer/eval/loc_agent_runner.py is agent-agnostic
(unchanged); examples import paths updated. claude_agent imports clean;
codex_agent needs the openai_codex SDK to run.
