+++ examples/codex_loc_agent.py
@@ -45,7 +45,7 @@ and ``usage``.
 import argparse
 import asyncio
 
-from codeminer.agent.codex_agent import CodexLocAgent
+from codeminer.clients.codex_agent import CodexLocAgent
 from codeminer.eval.loc_agent_runner import add_common_args, run_agent_baseline
 from codeminer.log_utils import get_logger
